modify triamgle function

+++ KELLY/triangle.py
@@ -1,8 +1,8 @@
 def triangle(bottom, height):
-    return float(bottom * height/2)
+    return bottom * height/2
 
 
 bottom = input("Please input bottom line length: ")
 height = input("Please input height: ")
 
-print(triangle(int(bottom), int(height)))
+print(triangle(float(bottom), float(height)))
